requirements.txt
anyio==3.6.2
fastapi==0.95.1
idna==3.4
joblib==1.2.0
numpy==1.24.3
pydantic==1.10.7
python-dotenv==1.0.0
scikit-learn==1.2.2
scipy==1.10.1
sniffio==1.3.0
starlette==0.26.1
threadpoolctl==3.1.0
typing_extensions==4.5.0
uvicorn==0.22.0
